Refactoring.
- Moved `UiSerializer`, `UiSerializable`, `UriTypeAdapter`, and `SerializableTypeAdapter` from `presentation` to `domain` package.
- Updated imports in `SuccessViewModel`, `ViewDocumentViewModel`, `SelectQtspViewModel`, `ViewDocumentUiConfig`, and `UiModule`.
- Marked `CertificateData` and `DocumentData` as `internal`.
- Removed unnecessary destination change listener in `RouterHost`.
- Added `@Suppress("DEPRECATION")` to `getStartingRoute` in `EudiRQESContainer`.
