+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/domain/di/UiModule.kt
@@ -18,8 +18,8 @@ package eu.europa.ec.rqesui.domain.di
 
 import eu.europa.ec.rqesui.presentation.navigation.RouterHost
 import eu.europa.ec.rqesui.presentation.navigation.RouterHostImpl
-import eu.europa.ec.rqesui.presentation.serializer.UiSerializer
-import eu.europa.ec.rqesui.presentation.serializer.UiSerializerImpl
+import eu.europa.ec.rqesui.domain.serializer.UiSerializer
+import eu.europa.ec.rqesui.domain.serializer.UiSerializerImpl
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.Single
