+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/infrastructure/config/data/DocumentData.kt
@@ -21,7 +21,7 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class DocumentData(
+internal data class DocumentData(
     val documentName: String,
     val uri: Uri,
 ) : Parcelable
