+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/presentation/extension/ContextExtensions.kt
@@ -99,33 +99,3 @@ internal suspend fun Context.loadPdf(
 
 private const val PDF_BITMAP_WIDTH = 1240
 private const val PDF_BITMAP_HEIGHT = 1754
-
-/**
- * Retrieves the file name from a given [Uri].
- *
- * This function attempts to extract the file name from the provided [Uri] using the content resolver.
- * If successful, it returns the file name.
- * If the file name cannot be determined from the [Uri], it returns the provided `fallbackName`.
- * If an exception occurs during the process, it also returns the `fallbackName`.
- *
- * @param context The [Context] used to access the content resolver.
- * @param fallbackName The name to return if the file name cannot be determined or an error occurs. Defaults to "Empty File Name".
- * @return The file name extracted from the [Uri], or the `fallbackName` if an error occurs or the name cannot be determined.
- */
-internal fun Uri.getFileName(
-    context: Context,
-    fallbackName: String = "Empty File Name"
-): String {
-    return try {
-        val fileName = context.contentResolver.query(this, null, null, null, null)?.use { cursor ->
-            val nameIndex = cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME)
-            cursor.moveToFirst()
-            cursor.getString(nameIndex)
-        }
-        //TODO what should be returned if fileName is null?
-        //return fileName.orEmpty()
-        fileName ?: fallbackName
-    } catch (e: Exception) {
-        fallbackName
-    }
-}
