+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/presentation/ui/container/EudiRQESContainer.kt
@@ -75,6 +75,7 @@ internal class EudiRQESContainer : ComponentActivity() {
 
     //TODO should we handle onNewIntent?
 
+    @Suppress("DEPRECATION")
     @Throws(EudiRQESUiError::class)
     private fun getStartingRoute(intent: Intent): String {
         val state = intent.getParcelableExtra<EudiRQESUi.State>(SDK_STATE)
