+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/presentation/ui/view_document/ViewDocumentViewModel.kt
@@ -23,7 +23,7 @@ import eu.europa.ec.rqesui.presentation.architecture.ViewEvent
 import eu.europa.ec.rqesui.presentation.architecture.ViewSideEffect
 import eu.europa.ec.rqesui.presentation.architecture.ViewState
 import eu.europa.ec.rqesui.presentation.entities.config.ViewDocumentUiConfig
-import eu.europa.ec.rqesui.presentation.serializer.UiSerializer
+import eu.europa.ec.rqesui.domain.serializer.UiSerializer
 import org.koin.android.annotation.KoinViewModel
 import org.koin.core.annotation.InjectedParam
